requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
jinja2>=3.1.0
pywin32>=306 ; sys_platform == 'win32'
pyinstaller>=6.0.0
pymupdf>=1.23.0
python-pptx>=0.6.21
